requirements.txt
mcp==1.27.0
python-dotenv==1.2.2
httpx==0.28.1
uvicorn==0.44.0
starlette==1.0.0
sse-starlette==3.3.4
